Fix schema errors

+++ src/utils/schema.cpp
@@ -113,7 +113,7 @@ SchemaKey schema::GetOffset(const char *className, uint32_t classKey, const char
 	return tableMap->Element(memberIndex);
 }
 
-void schema::NetworkStateChanged(int64 chainEntity, int64 nLocalOffset, int64 nArrayIndex)
+void schema::NetworkStateChanged(int64 chainEntity, uint32 nLocalOffset, int nArrayIndex)
 {
 	CNetworkVarChainer *chainEnt = reinterpret_cast<CNetworkVarChainer *>(chainEntity);
 	CEntityInstance *entity = chainEnt->GetObject();
